Refactorize tests
