+++ Netskope/tests/actions/test_action_replace_blocklist.py
@@ -4,20 +4,15 @@ import pytest
 import requests_mock
 from pydantic import ValidationError
 
-from netskope_modules import NetskopeModule
 from netskope_modules.actions.action_replace_blocklist import ReplaceBlocklistAction
 
 
 @pytest.fixture
-def replace_action(symphony_storage):
-    module = NetskopeModule()
-    action = ReplaceBlocklistAction(module=module, data_path=symphony_storage)
+def replace_action(symphony_storage, trigger):
+    trigger.module.configuration.base_url = "https://my.fake.netskope.com"
+    action = ReplaceBlocklistAction(module=trigger.module, data_path=symphony_storage)
     action.log = MagicMock()
     action.log_exception = MagicMock()
-    action.module.configuration = {
-        "base_url": "https://my.fake.netskope.com",
-        "api_token": "fake_api_token",
-    }
     return action
