+++ Netskope/tests/conftest.py
@@ -1,8 +1,13 @@
+import os
 from shutil import rmtree
 from tempfile import mkdtemp
+from unittest.mock import MagicMock
 
 import pytest
 from sekoia_automation import constants
+
+from netskope_modules import NetskopeModule
+from netskope_modules.connectors.connector_pull_events_v2 import NetskopeEventConnector
 
 
 @pytest.fixture
@@ -30,3 +35,43 @@ def symphony_storage():
 
     rmtree(constants.DATA_STORAGE)
     constants.DATA_STORAGE = original_storage
+
+
+@pytest.fixture
+def trigger(symphony_storage):
+    module = NetskopeModule()
+    module._trigger_configuration_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
+    connector = NetskopeEventConnector(module=module, data_path=symphony_storage)
+    # Avoid network access in unit tests.
+    connector.log = MagicMock()
+    connector.log_exception = MagicMock()
+    connector.push_events_to_intakes = MagicMock()
+    connector.module.configuration = {
+        "base_url": "https://my.fake.sekoia",
+        "api_token": "fake_api_token",
+    }
+    connector.configuration = {
+        "api_token": "api_token",
+        "intake_key": "intake_key",
+        "consumer_group": "",
+    }
+    return connector
+
+
+@pytest.fixture
+def integration_trigger(symphony_storage):
+    module = NetskopeModule()
+    module._community_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
+    connector = NetskopeEventConnector(module=module, data_path=symphony_storage)
+    connector.log = MagicMock()
+    connector.log_exception = MagicMock()
+    connector.push_events_to_intakes = MagicMock()
+    connector.module.configuration = {
+        "base_url": os.environ["NETSKOPE_BASE_URL"],
+        "api_token": os.environ["NETSKOPE_API_TOKEN"],
+    }
+    connector.configuration = {
+        "api_token": os.environ["NETSKOPE_API_TOKEN"],
+        "intake_key": "0123456789",
+    }
+    return connector
