+++ Netskope/tests/connectors/test_connector_pull_events_v2.py
@@ -8,31 +8,9 @@ import requests_mock
 from netskope_api.iterator.netskope_iterator import NetskopeIterator
 from sekoia_automation.exceptions import ModuleConfigurationError
 
-from netskope_modules import NetskopeModule
-from netskope_modules.connectors.connector_pull_events_v2 import NetskopeEventConnector, NetskopeEventConsumer
+from netskope_modules.connectors.connector_pull_events_v2 import NetskopeEventConsumer
 from netskope_modules.constants import MESSAGE_CANNOT_CONSUME_SERVICE
 from netskope_modules.types import NetskopeAlertType, NetskopeEventType
-
-
-@pytest.fixture
-def trigger(symphony_storage):
-    module = NetskopeModule()
-    module._trigger_configuration_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
-    trigger = NetskopeEventConnector(module=module, data_path=symphony_storage)
-    # mock the log function of trigger that requires network access to the api for reporting
-    trigger.log = MagicMock()
-    trigger.log_exception = MagicMock()
-    trigger.push_events_to_intakes = MagicMock()
-    trigger.module.configuration = {
-        "base_url": "https://my.fake.sekoia",
-        "api_token": "fake_api_token",
-    }
-    trigger.configuration = {
-        "api_token": "api_token",
-        "intake_key": "intake_key",
-        "consumer_group": "",
-    }
-    return trigger
 
 
 def test_user_agent(trigger):
@@ -240,6 +218,25 @@ def test_next_batch_no_consume_service(trigger):
         assert not mock_time.sleep.called
 
 
+def test_next_batch_invalid_api_token(trigger):
+    with patch("netskope_modules.connectors.connector_pull_events_v2.time") as mock_time:
+        iterator = trigger.create_iterator(NetskopeEventType.ALERT, NetskopeAlertType.DLP)
+        iterator.client.get = MagicMock(
+            side_effect=ValueError(
+                "Invalid API token TOKEN configured to access the endpoint https://my.fake.sekoia/api/v2/events/dataexport/alerts/dlp"
+            )
+        )
+
+        consumer = NetskopeEventConsumer(trigger, "alert-dlp", iterator)
+        consumer.stop = Mock()
+
+        consumer.next_batch()
+
+        assert trigger.push_events_to_intakes.call_count == 0
+        assert consumer.stop.called
+        assert not mock_time.sleep.called
+
+
 def test_next_batch_403_service(trigger):
     with (
         patch("netskope_modules.connectors.connector_pull_events_v2.time") as mock_time,
@@ -331,43 +328,15 @@ def test_stop_consumers(trigger):
     assert incident_consumer.stop.called
 
 
-def test_undefined_base_url_should_raise_exception(symphony_storage):
-    module = NetskopeModule()
-    module._community_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
-    trigger = NetskopeEventConnector(module=module, data_path=symphony_storage)
-    # mock the log function of trigger that requires network access to the api for reporting
-    trigger.log = MagicMock()
-    trigger.log_exception = MagicMock()
-    trigger.push_events_to_intakes = MagicMock()
-    trigger.module.configuration = {
-        "base_url": None,
-        "api_token": "fake_api_token",
-    }
-    trigger.configuration = {
-        "api_token": "api_token",
-        "intake_key": "intake_key",
-    }
+def test_undefined_base_url_should_raise_exception(trigger):
+    trigger.module.configuration.base_url = None
     with pytest.raises(ModuleConfigurationError):
         trigger.run()
 
 
 @pytest.mark.skipif("{'NETSKOPE_BASE_URL', 'NETSKOPE_API_TOKEN'}" ".issubset(os.environ.keys()) == False")
-def test_fetch_next_batch_integration(symphony_storage):
-    module = NetskopeModule()
-    module._community_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
-    trigger = NetskopeEventConnector(module=module, data_path=symphony_storage)
-    # mock the log function of trigger that requires network access to the api for reporting
-    trigger.log = MagicMock()
-    trigger.log_exception = MagicMock()
-    trigger.push_events_to_intakes = MagicMock()
-    trigger.module.configuration = {
-        "base_url": os.environ["NETSKOPE_BASE_URL"],
-        "api_token": os.environ["NETSKOPE_API_TOKEN"],
-    }
-    trigger.configuration = {
-        "api_token": os.environ["NETSKOPE_API_TOKEN"],
-        "intake_key": "0123456789",
-    }
+def test_fetch_next_batch_integration(integration_trigger):
+    trigger = integration_trigger
     iterator = trigger.create_iterator(NetskopeEventType.ALERT, NetskopeAlertType.DLP)
     consumer = NetskopeEventConsumer(trigger, "alert-dlp", iterator)
 
@@ -380,22 +349,8 @@ def test_fetch_next_batch_integration(symphony_storage):
 
 
 @pytest.mark.skipif("{'NETSKOPE_BASE_URL', 'NETSKOPE_API_TOKEN'}" ".issubset(os.environ.keys()) == False")
-def test_run_integration(symphony_storage):
-    module = NetskopeModule()
-    module._community_uuid = "ec92e51c-d45e-47b1-b820-29b97721623f"
-    trigger = NetskopeEventConnector(module=module, data_path=symphony_storage)
-    # mock the log function of trigger that requires network access to the api for reporting
-    trigger.log = MagicMock()
-    trigger.log_exception = MagicMock()
-    trigger.push_events_to_intakes = MagicMock()
-    trigger.module.configuration = {
-        "base_url": os.environ["NETSKOPE_BASE_URL"],
-        "api_token": os.environ["NETSKOPE_API_TOKEN"],
-    }
-    trigger.configuration = {
-        "api_token": os.environ["NETSKOPE_API_TOKEN"],
-        "intake_key": "0123456789",
-    }
+def test_run_integration(integration_trigger):
+    trigger = integration_trigger
     main_thread = Thread(target=trigger.run)
     main_thread.start()
